backend/requirements.txt
# Django core
Django>=5.1,<5.3
djangorestframework>=3.15
django-cors-headers>=4.6
django-filter>=24.3

# Auth
djangorestframework-simplejwt>=5.4

# Database
psycopg[binary]>=3.2
django-environ>=0.11

# WebSocket / Chat temps réel
channels>=4.2
daphne>=4.1

# Tâches asynchrones
celery>=5.4
django-celery-beat>=2.7

# Médias / Fichiers
Pillow>=11.0

# Paiement Orange Money
requests>=2.32
cryptography>=44.0

# Utils
django-extensions>=3.2
whitenoise>=6.8

# Dev
black>=24.0
ruff>=0.8
